ChefMate/requirements.txt
streamlit
pandas
pickle5
google-generativeai
Pillow